+++ components/src/components/md-toc-item/md-toc-item.tsx
@@ -1,4 +1,5 @@
 import { Component, Prop, h } from "@stencil/core";
+import { ReferenceType } from "../md-link/md-link-files";
 
 @Component({
   tag: "md-toc-item",
@@ -8,9 +9,30 @@ export class TableOfContentsItem {
   @Prop() chapter: string;
   @Prop() name: string = "";
   @Prop() indent: string;
+  @Prop() href: string;
+
+  connectedCallback() {
+    document.querySelectorAll("md-link-storage").forEach((item) => {
+      const storage = item;
+      storage.update(this.name, {
+        displayValue: "Chapter " + this.chapter,
+        type: ReferenceType.chapter,
+        href: "#" + this.href,
+        isUsed: false,
+        index: Number.MAX_VALUE
+      });
+
+      storage.update(this.href, {
+        displayValue: "Chapter " + this.chapter,
+        type: ReferenceType.chapter,
+        href: "#" + this.href,
+        isUsed: false,
+        index: Number.MAX_VALUE
+      });
+    });
+  }
 
   render() {
-    const href = "#" + (this.name).toLowerCase().replace(new RegExp(' ', 'g'), '-');
-    return <a class={"md-toc-item md-toc-item-indent-" + this.indent} href={href}><span class="md-toc-item-chapter">{this.chapter}</span><span class="md-toc-item-name">{this.name}</span></a>;
+    return <a class={"md-toc-item md-toc-item-indent-" + this.indent} href={"#" + this.href}><span class="md-toc-item-chapter">{this.chapter}</span><span class="md-toc-item-name">{this.name}</span></a>;
   }
 }